+++ R/creel-estimates-total-catch.R
@@ -273,7 +273,7 @@ estimate_total_catch <- function(
       product_variance = product_variance,
       ci_type = ci_type
     )
-    return(new_creel_estimates(
+    return(new_creel_estimates( # nolint: object_usage_linter
       # nolint: object_usage_linter
       estimates = tibble::as_tibble(estimates_df),
       method = "product-total-catch",
@@ -281,7 +281,8 @@ estimate_total_catch <- function(
       design = design,
       conf_level = conf_level,
       by_vars = by_info$all_vars,
-      effort_target = target
+      effort_target = target,
+      unit = "fish"
     ))
   }
 
@@ -786,6 +787,7 @@ estimate_total_catch_sections <- function(
     design = design,
     conf_level = conf_level,
     by_vars = if (!is.null(by_vars)) c("section", by_vars) else "section",
-    effort_target = target
+    effort_target = target,
+    unit = "fish"
   )
 }