+++ R/creel-estimates-total-harvest.R
@@ -219,7 +219,7 @@ estimate_total_harvest <- function(
       target = target,
       product_variance = product_variance
     )
-    return(new_creel_estimates(
+    return(new_creel_estimates( # nolint: object_usage_linter
       # nolint: object_usage_linter
       estimates = tibble::as_tibble(estimates_df),
       method = "product-total-harvest",
@@ -227,7 +227,8 @@ estimate_total_harvest <- function(
       design = design,
       conf_level = conf_level,
       by_vars = by_info$all_vars,
-      effort_target = target
+      effort_target = target,
+      unit = "fish"
     ))
   }
 
@@ -659,6 +660,7 @@ estimate_total_harvest_sections <- function(
     design = design,
     conf_level = conf_level,
     by_vars = if (!is.null(by_vars)) c("section", by_vars) else "section",
-    effort_target = target
+    effort_target = target,
+    unit = "fish"
   )
 }